fix : parameters moved to params.py
