+++ mytimer/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 """mytimer modules."""
-from mytimer.functions import MY_TIMER_VERSION
+from mytimer.params import MY_TIMER_VERSION
 __version__ = MY_TIMER_VERSION
