+++ mytimer/__main__.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 """mytimer main."""
-from mytimer.functions import countdown_timer, countup_timer, check_null_time, MY_TIMER_VERSION, FACES_MAP
+from mytimer.params import MY_TIMER_VERSION, FACES_MAP
+from mytimer.functions import countdown_timer, countup_timer, check_null_time
 import argparse
