+++ mytimer/functions.py
@@ -3,30 +3,8 @@
 import os
 import sys
 import time
+from mytimer.params import *
 from art import tprint
-
-MY_TIMER_VERSION = "0.3"
-WRONG_INPUT_ERROR = "[Error] Wrong input"
-SOUND_ERROR_MESSAGE = "[Error] Unable to play sound"
-INPUT_EXAMPLE = "Example: python -m mytimer --hour=1 --minute=1 --second=1"
-FACES_MAP = {
-    1: 'bulbhead',
-    2: 'soft',
-    3: '4max',
-    4: '5x7',
-    5: 'charact4',
-    6: 'o8',
-    7: 'alphabet',
-    8: 'shadow',
-    9: 'speed',
-    10: 'rounded',
-    11: 'chartri',
-    12: 'standard',
-    13: 'contessa',
-    14: 'avatar',
-    15: 'mini',
-    16: 'twopoint'}
-
 
 def check_null_time(args):
     """
@@ -36,8 +14,7 @@ def check_null_time(args):
     :type args: argparse.Namespace
     :return: result as bool
     """
-    time_elements = ["minute", "second", "hour"]
-    for item in time_elements:
+    for item in TIME_ELEMENTS:
         if getattr(args, item) is not None:
             return False
     return True
